add: test for app, middleware for controllers, new controller

+++ app/Http/Controllers/MovieController.php
@@ -15,19 +15,22 @@ use function PHPUnit\Framework\isEmpty;
 
 class MovieController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except('showMovies', 'showMovie');
+    }
+
     public function showMovies()
     {
         $user = auth()->user();
-        $movies = Movie::where('status','verified')->get();
+        $movies = Movie::where('status', 'verified')->get();
         $movies->each(function ($movie) {
             $movie->release_date = Carbon::parse($movie->release_date)->format('d.m.Y');
         });
-        foreach ($movies as $movie)
-        {
+        foreach ($movies as $movie) {
             $averageRating = Rate::where('movie_id', $movie->id)->avg('rate');
 
-            if ($user)
-            {
+            if ($user) {
                 $fav = Collection::where([
                     'movie_id' => $movie->id,
                     'user_id' => $user->id,
@@ -55,18 +58,18 @@ class MovieController extends Controller
     public function addMovieForm()
     {
         $artist = Artist::all();
-        return view("addMovie",compact('artist'));
+        return view("addMovie", compact('artist'));
     }
 
     public function changeTrailerLink($link)
     {
-        if (str_contains($link, "embed"))
-        {
+        if (str_contains($link, "embed")) {
             return $link;
         }
 
         return str_replace("watch?v=", "embed/", $link);
     }
+
     public function changeSoundtrackLink($link)
     {
         if (str_contains($link, "embed")) {
@@ -81,7 +84,7 @@ class MovieController extends Controller
             }
 
             $middle = "/embed";
-            return $prefix.$middle.$suffix;
+            return $prefix . $middle . $suffix;
         }
     }
 
@@ -100,18 +103,19 @@ class MovieController extends Controller
             }
         }];
     }
+
     protected function validator(array $data)
     {
         return Validator::make($data, array_merge([
             'title' => ['required', 'string'],
             'releaseDate' => ['required', 'date'],
             'description' => ['required', 'string'],
-            'trailerLink' => ['required', 'string', 'starts_with:https://www.youtube.com/watch?v=', 'max:43','regex:/[A-Za-z0-9]{11}$/'],
+            'trailerLink' => ['required', 'string', 'starts_with:https://www.youtube.com/watch?v=', 'max:43', 'regex:/[A-Za-z0-9]{11}$/'],
             'soundtrackLink' => ['required', 'string', 'regex:/https:\/\/open\.spotify\.com\/album\/[A-Za-z0-9]{22}\?si=[A-Za-z0-9_-]+$/'],
             'artists' => ['array', 'min:1', 'exists:artists,id'],
-            'poster' => ['image', 'required', 'mimes:jpeg,png,jpg', 'mimetypes:image/jpeg,image/png,image/jpg','max: 16777215',$this->checkImage()],
+            'poster' => ['image', 'required', 'mimes:jpeg,png,jpg', 'mimetypes:image/jpeg,image/png,image/jpg', 'max: 16777215', $this->checkImage()],
             'images' => ['required'],
-            'images.*' => ['image', 'mimes:jpeg,png,jpg', 'mimetypes:image/jpeg,image/png,image/jpg','max: 16777215', $this->checkImage()],
+            'images.*' => ['image', 'mimes:jpeg,png,jpg', 'mimetypes:image/jpeg,image/png,image/jpg', 'max: 16777215', $this->checkImage()],
         ]));
     }
 
@@ -128,19 +132,8 @@ class MovieController extends Controller
 
     public function isDuplicateForAdd($movie)
     {
-        $duplicate = Movie::where(['title'=> $movie->title])->where(['release_date' => $movie->releaseDate])->get();
-        if(sizeof($duplicate)>0)
-        {
-            return true;
-        }
-        return false;
-    }
-
-    public function isDuplicateForEdit($movie)
-    {
-        $duplicate = Movie::where(['title'=> $movie->title])->where(['release_date' => $movie->releaseDate])->get();
-        if(sizeof($duplicate)>1)
-        {
+        $duplicate = Movie::where(['title' => $movie->title])->where(['release_date' => $movie->releaseDate])->get();
+        if (sizeof($duplicate) > 0) {
             return true;
         }
         return false;
@@ -148,7 +141,6 @@ class MovieController extends Controller
 
     public function addMovie(Request $request)
     {
-        $this->middleware('auth');
         $validator = $this->validator($request->all());
 
         if ($validator->fails()) {
@@ -158,9 +150,9 @@ class MovieController extends Controller
         }
 
         $isDuplicate = $this->isDuplicateForAdd($request);
-        if(!$isDuplicate) {
+        if (!$isDuplicate) {
             $userId = auth()->user()->id;
-            $p = $this->resizeImage($request->poster,300);
+            $p = $this->resizeImage($request->poster, 300);
             $t = $this->changeTrailerLink($request->trailerLink);
             $s = $this->changeSoundtrackLink($request->soundtrackLink);
 
@@ -186,46 +178,46 @@ class MovieController extends Controller
 
             $movie->artist()->sync($request->artists);
             return redirect('/movies')->with('success', 'Film został dodany pomyślnie.');
-        }
-        else
-        {
+        } else {
             return redirect('/movies')->with('error', 'Film już istnieje w bazie danych!');
         }
 
     }
+
     public function showMovie($id)
     {
         $userRate = null;
-        if(auth()->user() != null)
-        {
+        if (auth()->user() != null) {
             $user_id = auth()->user()->id;
             $userRate = Rate::where('user_id', $user_id)->where('movie_id', $id)->first();
         }
         $movie = Movie::find($id);
         $movie->release_date = Carbon::parse($movie->release_date)->format('d.m.Y');
         $avgRate = $this->avgMovie($id);
-        return view("showMovie", compact("movie","userRate","avgRate"));
+        return view("showMovie", compact("movie", "userRate", "avgRate"));
     }
 
     public function deleteMovie($id)
     {
         $toDelete = Movie::find($id);
-        if($toDelete)
-        {
-            $toDelete->delete();
-            return redirect()->route('showMovies')->with('success', 'Film został usunięty pomyślnie.');
-        }
-        else
-        {
-            return redirect()->route('showMovies')->with('error', "Film o podanym ID nie istenieje!");
+        if (auth()->user()->id == $toDelete->user_id or auth()->user()->role == 'admin') {
+            if ($toDelete) {
+                $toDelete->delete();
+                return redirect()->route('showMovies')->with('success', 'Film został usunięty pomyślnie.');
+            } else {
+                return redirect()->route('showMovies')->with('error', "Film o podanym ID nie istenieje!");
 
+            }
         }
+        return redirect()->route('showMovies')->with('error', "Wybrany film nie należy do Ciebie!");
+
     }
+
     public function editMovieForm($id)
     {
         $movie = Movie::find($id);
         $artist = Artist::all();
-        return view("editMovie", compact("movie","artist"));
+        return view("editMovie", compact("movie", "artist"));
     }
 
     protected function validatorForEdit(array $data)
@@ -237,8 +229,8 @@ class MovieController extends Controller
             'trailerLink' => ['required', 'string', 'max:255', 'regex:/^(?:https?:\/\/)?(?:www\.)?(?:youtube\.com\/(?:[^\/\n\s]+\/\S+\/|(?:v|e(?:mbed)?)\/|\S*?[?&]v=))([a-zA-Z0-9_-]{11})\b/'],
             'soundtrackLink' => ['required', 'string', 'max:255', 'regex:/^https:\/\/open\.spotify\.com\/album\/[a-zA-Z0-9]{22}(\?si=[A-Za-z0-9_-]+)?$/'],
             'artists' => ['array', 'min:1', 'exists:artists,id'],
-            'poster' => ['image', 'mimes:jpeg,png,jpg', 'mimetypes:image/jpeg,image/png,image/jpg','max: 16777215',$this->checkImage()],
-            'images.*' => ['image', 'mimes:jpeg,png,jpg', 'mimetypes:image/jpeg,image/png,image/jpg','max: 16777215',$this->checkImage()],
+            'poster' => ['image', 'mimes:jpeg,png,jpg', 'mimetypes:image/jpeg,image/png,image/jpg', 'max: 16777215', $this->checkImage()],
+            'images.*' => ['image', 'mimes:jpeg,png,jpg', 'mimetypes:image/jpeg,image/png,image/jpg', 'max: 16777215', $this->checkImage()],
             'imagesToDelete' => ['array', function ($attribute, $value, $fail) use ($data) {
                 if (count($value) === count($data['images']) && empty($data['images'])) {
                     $fail("Film musi zawierać przynajmniej jeden fotos!");
@@ -246,54 +238,55 @@ class MovieController extends Controller
             }]
         ]));
     }
-    public function editMovie(Request $request,$id)
+
+    public function editMovie(Request $request, $id)
     {
         $edit = Movie::find($id);
-        if($request->poster==null)
-        {
-               $request->poster = $edit->poster;
-        }
-
-        if ($request->imagesToDelete != null)
-        {
-            $imagesToDelete = $request->input('imagesToDelete');
-            $edit->image()->whereIn('id', $imagesToDelete)->delete();
-        }
-        $t = $this->changeTrailerLink($request->trailerLink);
-        $s = $this->changeSoundtrackLink($request->soundtrackLink);
-        $request->trailer_link = $t;
-        $request->soundtrackLink = $s;
-
-        $validator = $this->validatorForEdit($request->all());
-        if ($validator->fails()) {
-            return redirect()->back()
-                ->withErrors($validator)
-                ->withInput();
-        }
-        $p = $this->resizeImage($request->poster,300);
-
-        $edit->update([
-            'title' => $request->title,
-            'release_date' => $request->releaseDate,
-            'description' => $request->description,
-            'trailer_link' => $t,
-            'soundtrack_link' => $s,
-            'poster' => $p,
-            'status' => 'pending',
-        ]);
-        if (!empty($request->images)) {
-            foreach ($request->images as $image) {
-                $i = $this->resizeImage($image,600);
-                $edit->image()->create([
-                    'image' => $i,
-                ]);
+        if (auth()->user()->id == $edit->user_id or auth()->user()->role == 'admin') {
+            if ($request->poster == null) {
+                $request->poster = $edit->poster;
             }
+
+            if ($request->imagesToDelete != null) {
+                $imagesToDelete = $request->input('imagesToDelete');
+                $edit->image()->whereIn('id', $imagesToDelete)->delete();
+            }
+            $t = $this->changeTrailerLink($request->trailerLink);
+            $s = $this->changeSoundtrackLink($request->soundtrackLink);
+            $request->trailer_link = $t;
+            $request->soundtrackLink = $s;
+
+            $validator = $this->validatorForEdit($request->all());
+            if ($validator->fails()) {
+                return redirect()->back()
+                    ->withErrors($validator)
+                    ->withInput();
+            }
+            $p = $this->resizeImage($request->poster, 300);
+
+            $edit->update([
+                'title' => $request->title,
+                'release_date' => $request->releaseDate,
+                'description' => $request->description,
+                'trailer_link' => $t,
+                'soundtrack_link' => $s,
+                'poster' => $p,
+                'status' => 'pending',
+            ]);
+            if (!empty($request->images)) {
+                foreach ($request->images as $image) {
+                    $i = $this->resizeImage($image, 600);
+                    $edit->image()->create([
+                        'image' => $i,
+                    ]);
+                }
+            }
+            if (auth()->user()->role == 'admin') {
+                return redirect()->route('showAdminPanel')->with('success', 'Film został zaktualizowany pomyślnie.');
+            }
+            return redirect()->route('showMovie', ['id' => $id])->with('success', 'Film został zaktualizowany pomyślnie.');
         }
-        if(auth()->user()->role=='admin')
-        {
-            return redirect()->route('showAdminPanel')->with('success', 'Film został zaktualizowany pomyślnie.');
-        }
-        return redirect()->route('showMovie',['id'=>$id])->with('success', 'Film został zaktualizowany pomyślnie.');
+
     }
 
     public function avgMovie($id)
@@ -311,56 +304,49 @@ class MovieController extends Controller
     public function addToCollection(Request $request)
     {
         $user = auth()->user();
+        $movie = Movie::find($request->id);
 
-        if ($user != null) {
-            $movie = Movie::find($request->id);
+        if ($movie != null) {
+            $existingRecord = Collection::where([
+                'movie_id' => $request->id,
+                'user_id' => $user->id,
+                'name' => $request->name,
+            ])->first();
 
-            if ($movie != null) {
-                $existingRecord = Collection::where([
-                    'movie_id' => $request->id,
-                    'user_id' => $user->id,
-                    'name' => $request->name,
-                ])->first();
-
-                if ($existingRecord) {
-                    return redirect()->route('showMovies')->with('error', "Film już istnieje w kolecji");
-                }
-
-                $validator = $this->validatorCollection($request->all());
-                if ($validator->fails()) {
-                    return redirect()->back()
-                        ->withErrors($validator)
-                        ->withInput();
-                }
-
-                $collection = Collection::create([
-                    'user_id' => $user->id,
-                    'movie_id' => $request->id,
-                    'name' => $request->name,
-                ]);
-
-                if (!$collection) {
-                    return redirect()->route('showMovies')->with('error', 'Coś poszło nie tak. Spróbuj ponownie póżniej.');
-                }
-                return redirect()->route('showMovies')->with('success', 'Film został dodany pomyślnie do kolekcji.');
+            if ($existingRecord) {
+                return redirect()->route('showMovies')->with('error', "Film już istnieje w kolecji");
             }
-        }
 
-        return redirect()->route('showMovies')->with('error', 'Dostęp tylko dla zalogowanych!');
+            $validator = $this->validatorCollection($request->all());
+            if ($validator->fails()) {
+                return redirect()->back()
+                    ->withErrors($validator)
+                    ->withInput();
+            }
+
+            $collection = Collection::create([
+                'user_id' => $user->id,
+                'movie_id' => $request->id,
+                'name' => $request->name,
+            ]);
+
+            if (!$collection) {
+                return redirect()->route('showMovies')->with('error', 'Coś poszło nie tak. Spróbuj ponownie póżniej.');
+            }
+            return redirect()->route('showMovies')->with('success', 'Film został dodany pomyślnie do kolekcji.');
+        }
     }
+
     public function deleteFromCollection(Request $request)
     {
         $user_id = auth()->user()->id;
-        $toDelete = Collection::where(['user_id' => $user_id])->where(['name'=>$request->name])->where(['movie_id' => $request->id]);
-        if($toDelete)
-        {
+        $toDelete = Collection::where(['user_id' => $user_id])->where(['name' => $request->name])->where(['movie_id' => $request->id]);
+
+        if ($toDelete) {
             $toDelete->delete();
             return redirect()->route('showMovies')->with('success', "Film został usunięty z kolekcji.");
-        }
-        else
-        {
+        } else {
             return redirect()->route('showMovies')->with('error', "Coś poszło nie tak. Spróbuj ponownie.");
-
         }
     }
